requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
databases==0.8.0
python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.23.4
pydantic==2.5.2
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary  # for PostgreSQL